Functions

Functions exported from Batsrus:

Batsrus.convertBox2VTK — Method

convertBoxVTK(filename; dir=".", gridType=1, verbose=false)

Convert 3D structured Tecplot data to VTK.

Batsrus.convertVTK — Function

convertVTK(head, data, connectivity, filename="out")

Convert unstructured Tecplot data to VTK. Note that if using voxel type data in VTK, the connectivity sequence is different from Tecplot. Note that the 3D connectivity sequence in Tecplot is the same with the hexahedron type in VTK, but different with the voxel type. The 2D connectivity sequence is the same as the quad type, but different with the pixel type. For example, in 3D the index conversion is:

# PLT to VTK voxel index_ = [1 2 4 3 5 6 8 7]
for i = 1:2
   connectivity = swaprows!(connectivity, 4*i-1, 4*i)
end

Batsrus.cutdata — Method

cutdata(data, var; plotrange=[-Inf,Inf,-Inf,Inf], cut=' ',
	cutPlaneIndex=1)

Get 2D plane cut data of 3D box data.

Batsrus.readdata — Method

readdata(filenameIn, (, dir=".", npict=1, verbose=false))

Read data from BATSRUS output files. Stores the npict snapshot from an ascii or binary data file into the coordinates x and data w arrays. Filenames can be provided with wildcards.

Examples

filename = "1d_raw*"
data = readdata(filename)

Batsrus.readlogdata — Method

Read information from log file.

Batsrus.readtecdata — Method

readtecdata(filename, verbose=false)

Return header, data and connectivity from BATSRUS Tecplot outputs. Both 2D and 3D binary and ASCII formats are supported.

Examples

filename = "3d_ascii.dat"
head, data, connectivity = readtecdata(filename)

Batsrus.showhead — Method

showhead(file, filehead)

Displaying file header information.

Batsrus.showhead — Method

showhead(data)

Displaying file information for the Data type.

Batsrus.subsurface — Method

subsurface(x, y, data, limits)
subsurface(x, y, u, v, limits)

Extract subset of 2D surface dataset. This is a simplified version of subvolume.

Batsrus.subvolume — Method

subvolume(x, y, z, data, limits)
subvolume(x, y, z, u, v, w, limits)

Extract subset of 3D dataset in ndgrid format.