Functions

Functions exported from SWMF:

SWMF.convertVTK — Function

convertVTK(head, data, connectivity, filename)

Convert 3D unstructured Tecplot data to VTK. Note that if using voxel type data in VTK, the connectivity sequence is different from Tecplot.

SWMF.readdata — Method

readdata(filenames,(, dir=".", npict=1, verbose=false))

Read data from BATSRUS output files. Stores the npict-th snapshot from an ascii or binary data file into the x [coordinates] and w [data] arrays. Filenames can be provided with wildcards.

Examples

filenames = "1d_raw*"
head, data, list = readdata(filenames)

SWMF.readlogdata — Method

readlogdata(filename)

Read information from log file.

SWMF.readtecdata — Method

readtecdata(filename, IsBinary=false, verbose=false)

Return header, data and connectivity from BATSRUS Tecplot outputs. Both binary and ascii formats are supported.

Examples

filename = "3d_ascii.dat"
head, data, connectivity = readtecdata(filename)