Functions

Functions exported from SWMF:

SWMF.convertVTK — Function

convertVTK(head, data, connectivity, filename)

Convert 3D unstructured Tecplot data to VTK. Note that if using voxel type data in VTK, the connectivity sequence is different from Tecplot.

SWMF.cutdata — Method

cutdata(data, var; plotrange=[-Inf,Inf,-Inf,Inf], cut=' ',
	cutPlaneIndex=1)

Get 2D plane cut data of 3D box data.

SWMF.readdata — Method

readdata(filenameIn, (, dir=".", npict=1, verbose=false))

Read data from BATSRUS output files. Stores the npict snapshot from an ascii or binary data file into the coordinates x and data w arrays. Filenames can be provided with wildcards.

Examples

filename = "1d_raw*"
data = readdata(filename)

SWMF.readlogdata — Method

Read information from log file.

SWMF.readtecdata — Method

readtecdata(filename, IsBinary=false, verbose=false)

Return header, data and connectivity from BATSRUS Tecplot outputs. Both binary and ASCII formats are supported. The default is reading pure ASCII data.

Examples

filename = "3d_ascii.dat"
head, data, connectivity = readtecdata(filename)

SWMF.subsurface — Method

subsurface(x, y, data, limits)
subsurface(x, y, u, v, limits)

Extract subset of 2D surface dataset. This is a simplified version of subvolume.

SWMF.subvolume — Method

subvolume(x, y, z, data, limits)
subvolume(x, y, z, u, v, w, limits)

Extract subset of 3D dataset in ndgrid format.