APIs

Public

Types

Batsrus.Data — Type

Primary data storage type

Functions

Batsrus.convertIDLtoVTK — Method

convertIDLtoVTK(filename; dir=".", gridType=1, verbose=false)

Convert 3D BATSRUS *.out to VTK. If gridType==1, it converts to the rectilinear grid; if gridType==2, it converts to the structured grid. If filename does not end with "out", it tries to find the ".info" and ".tree" file with the same name tag and generates 3D unstructured VTU file.

Batsrus.convertTECtoVTU — Function

convertTECtoVTU(head, data, connectivity, filename="out")

Convert unstructured Tecplot data to VTK. Note that if using voxel type data in VTK, the connectivity sequence is different from Tecplot. Note that the 3D connectivity sequence in Tecplot is the same with the hexahedron type in VTK, but different with the voxel type. The 2D connectivity sequence is the same as the quad type, but different with the pixel type. For example, in 3D the index conversion is:

# PLT to VTK voxel index_ = [1 2 4 3 5 6 8 7]
for i = 1:2
   connectivity = swaprows!(connectivity, 4*i-1, 4*i)
end

Batsrus.cutdata — Method

cutdata(data, var; plotrange=[-Inf,Inf,-Inf,Inf], dir="x", sequence=1)

Get 2D plane cut in orientation dir for var out of 3D box data within plotrange. The returned 2D data lies in the sequence plane from - to + in dir.

Batsrus.cutplot — Method

cutplot(data, var; plotrange=[-Inf,Inf,-Inf,Inf], dir="x", plotinterval=0.1,
   density=1.0, sequence=1, level=20)

2D plane cut contourf of 3D box data.

Batsrus.getConnectivity — Method

Get cell connectivity list.

Batsrus.getvar — Method

Return variable data from string var.

Batsrus.getvars — Method

getvars(data::Data, Names::Vector) -> Dict

Return variables' data as a dictionary from string vector. See also: getvar.

Batsrus.plotdata — Method

plotdata(data, func, (...))

Plot the variable from SWMF output.

plotdata(data, "p", plotmode="contbar")

plotdata(data, "p", plotmode="grid")

plotdata(data, func, plotmode="trimesh",plotrange=plotrange, plotinterval=0.2)

Input arguments

data::Data: output data.

func::String: variables for plotting.

plotmode::String: (optional) type of plotting ["cont","contbar"]...

plotrange::Vector: (optional) range of plotting.

plotinterval: (optional) interval for interpolation.

level: (optional) level of contour.

dir: (optional) 2D cut plane orientation from 3D outputs ["x","y","z"].

sequence: (optional) sequence of plane from - to + in that direction.

multifigure: (optional) 1 for multifigure display, 0 for subplots.

verbose: (optional) display additional information.

density: (optional) density for streamlines.

Right now this can only deal with 2D plots or 3D cuts. Full 3D plots may be supported in the future.

Batsrus.plotlogdata — Method

plotlogdata(data, head, func; plotmode="line")

Plot information from log file.

Input arguments

data::Array: output data.

head::NamedTuple: header info.

func::String: variables for plotting.

plotmode::String: type of plotting ["line","scatter"].

Batsrus.readdata — Method

readdata(filenameIn; dir=".", npict=1, verbose=false)

Read data from BATSRUS output files. Stores the npict snapshot from an ascii or binary data file into the arrays of coordinates x and data w. Filename can be provided with wildcards.

Examples

data = readdata("1d_raw*")

Batsrus.readhead — Method

Return header from info file. Currently only designed for 2D and 3D.

Batsrus.readlogdata — Method

Read information from log file.

Batsrus.readtecdata — Method

readtecdata(filename; verbose=false)

Return header, data and connectivity from BATSRUS Tecplot outputs. Both 2D and 3D binary and ASCII formats are supported.

Examples

filename = "3d_ascii.dat"
head, data, connectivity = readtecdata(filename)

Batsrus.readtree — Method

Return BATL tree structure.

Batsrus.showhead — Function

showhead(file, filehead)

Displaying file header information.

Batsrus.showhead — Method

showhead(data)

showhead(io, data)

Display file information of data.

Batsrus.streamslice — Method

streamslice(data::Data, var::String; plotrange=[-Inf,Inf,-Inf,Inf], dir="x",
   plotinterval=0.1, density=1.0, sequence=1, color="w", linewidth=1.0)

Plot streamlines on 2D slices of 3D box data. Variable names in var string must be separated with ;.

Batsrus.subsurface — Method

subsurface(x, y, data, limits)
subsurface(x, y, u, v, limits)

Extract subset of 2D surface dataset. See also: subvolume.

Batsrus.subvolume — Method

subvolume(x, y, z, data, limits)
subvolume(x, y, z, u, v, w, limits)

Extract subset of 3D dataset in ndgrid format. See also: subsurface.

PyPlot.contour — Function

contour(data, var, levels=0; plotrange, plotinterval, kwargs...)

Wrapper over the contour function in matplotlib.

PyPlot.contourf — Function

contourf(data, var, levels=0; plotrange, plotinterval, kwargs...)

Wrapper over the contourf function in matplotlib.

PyPlot.plot — Method

plot(data, var; kwargs...)

Wrapper over the plot function in matplotlib.

PyPlot.plot_surface — Method

plot_surface(data, var; plotrange, plotinterval, kwargs...)

Wrapper over the plot_surface function in matplotlib.

PyPlot.plot_trisurf — Method

plot_trisurf(data::Data, var::String;
   plotrange::Vector{Float64}=[-Inf,Inf,-Inf,Inf], kwargs...)

Wrapper over the plot_trisurf function in matplotlib.

PyPlot.scatter — Method

scatter(data, var; kwargs...)

Wrapper over the scatter function in matplotlib.

PyPlot.streamplot — Method

streamplot(data, var; plotrange, plotinterval=0.1, kwargs...)

Wrapper over the streamplot function in matplotlib. Streamplot does not have **kwargs in the API, but it supports density, color, and some other keywords.

PyPlot.tricontourf — Method

tricontourf(data, var; plotrange, plotinterval, kwargs...)

Wrapper over the tricontourf function in matplotlib.

Private

```@autodocs Modules = [Batsrus] Public = false